Develop/hw 8

+++ app/src/main/java/ru/internetcloud/strava/data/training/repository/TrainingRepositoryImpl.kt
@@ -1,39 +1,55 @@
 package ru.internetcloud.strava.data.training.repository
 
-import ru.internetcloud.strava.data.training.cache.datasource.TrainingLocalDataSourceImpl
+import ru.internetcloud.strava.data.training.cache.datasource.TrainingLocalDataSource
 import ru.internetcloud.strava.data.training.mapper.TrainingListItemMapper
 import ru.internetcloud.strava.data.training.mapper.TrainingMapper
-import ru.internetcloud.strava.data.training.network.datasource.TrainingRemoteApiDataSourceImpl
+import ru.internetcloud.strava.data.training.network.datasource.TrainingRemoteApiDataSource
+import ru.internetcloud.strava.domain.common.list.mvi.model.ListLoadParams
+import ru.internetcloud.strava.domain.common.list.mvi.model.ListState
 import ru.internetcloud.strava.domain.common.model.DataResponse
 import ru.internetcloud.strava.domain.common.model.Source
 import ru.internetcloud.strava.domain.training.TrainingRepository
 import ru.internetcloud.strava.domain.training.model.Training
 import ru.internetcloud.strava.domain.training.model.TrainingListItem
 
-class TrainingRepositoryImpl : TrainingRepository {
+class TrainingRepositoryImpl(
+    private val trainingRemoteApiDataSource: TrainingRemoteApiDataSource,
+    private val trainingLocalDataSource: TrainingLocalDataSource,
+    private val trainingMapper: TrainingMapper,
+    private val trainingListItemMapper: TrainingListItemMapper
+) : TrainingRepository {
 
-    private val trainingRemoteApiDataSource = TrainingRemoteApiDataSourceImpl()
-    private val trainingLocalDataSource = TrainingLocalDataSourceImpl()
-
-    private val trainingMapper = TrainingMapper()
-    private val trainingListItemMapper = TrainingListItemMapper()
-
-    override suspend fun getTrainings(page: Int): DataResponse<List<TrainingListItem>> {
-        var result = trainingRemoteApiDataSource.getTrainings(page)
+    override suspend fun getTrainings(params: ListLoadParams<Unit>): DataResponse<ListState<TrainingListItem>> {
+        val result = trainingRemoteApiDataSource.getTrainings(params)
 
         if (result is DataResponse.Success) {
-            trainingLocalDataSource.deleteAllTrainingListItems()
-            trainingLocalDataSource.deleteAllTrainings()
+            if (params.page == 1) {
+                // при загрузке первой страницы надо очищать полностью данные в БД,
+                // иначе удаленная тренировка зависнет в локальном кеше!
+                trainingLocalDataSource.deleteAllTrainingListItems()
+                trainingLocalDataSource.deleteAllTrainings()
+            }
             trainingLocalDataSource.insertTrainingListItems(
                 trainingListItemMapper.fromListDomainToListDbModel(
-                    result.data
+                    result.data.items
                 )
             )
-        } else {
+        }
+
+        return result
+    }
+
+    override suspend fun getTrainingsWithLocalCache(params: ListLoadParams<Unit>): DataResponse<ListState<TrainingListItem>> {
+        var result = getTrainings(params)
+
+        if (result is DataResponse.Error) {
             val listDbModel = trainingLocalDataSource.getTrainingListItems()
             if (listDbModel.isNotEmpty()) {
                 result = DataResponse.Success(
-                    data = trainingListItemMapper.fromListDbModelToListDomain(listDbModel),
+                    data = ListState(
+                        items = trainingListItemMapper.fromListDbModelToListDomain(listDbModel),
+                        isLastPage = true
+                    ),
                     source = Source.LocalCache
                 )
             }
